+++ di-monad/default.nix
@@ -1,14 +1,15 @@
-{ mkDerivation, base, containers, di-core, exceptions, lib, mtl
-, pipes, pipes-safe, primitive, stm, streaming, transformers
-, unliftio-core
+{ mkDerivation, base, containers, di-core, exceptions, lib
+, monad-control, mtl, pipes, pipes-safe, primitive, stm, streaming
+, transformers, transformers-base, unliftio-core
 }:
 mkDerivation {
   pname = "di-monad";
   version = "1.3.3";
   src = ./.;
   libraryHaskellDepends = [
-    base containers di-core exceptions mtl pipes pipes-safe primitive
-    stm streaming transformers unliftio-core
+    base containers di-core exceptions monad-control mtl pipes
+    pipes-safe primitive stm streaming transformers transformers-base
+    unliftio-core
   ];
   homepage = "https://github.com/k0001/di";
   description = "mtl flavoured typeful hierarchical structured logging for di-core";